Handling Ability Inputs
PlayerControls now includes a function/section for handling ability inputs, including a 'preparing' state, toggling of visual indicators that flash (animated), and processing directional inputs. If an adjacent object is detected, then it is marked as quantum. A system was created for implementing future dinosaur type abilities through a basic switch case in the TryAbility function