Working Quantum Mechanic
ObjectState now handles updating object sprites to match their actual type. In addition, a static list of marked quantum objects was added alongside a static function that will shuffle the object types among the hidden quantum objects. This function is called on a recurring timer on a new GameTimer script located on the main panel (this will also eventually handle syncing 2-frame animation times). In addition, these changes required the addition of another VisibilityCheck function (for generic game objects), slightly tweaking UndoObject, PanelStats, and PanelControls. A polish change of making a panel's border and ground tiles slightly transparent when in the process of dragging compliments the quantum mechanic well to indicate when a position is locked in