Dinosaur Swapping
Added dino swapping controls to PlayerControls and reworked storage of current type to instead index to a arrays for DInosaur type and remaining ability charges. Charges may be limited or unlimited for a particular stage. Also created PlayerSpriteSwapper for handling the effects of shrinking, swapping, and unshrinking the player upon swapping dino type