Object Data Restructure
Created serializable ObjectData struct  that stores all object data (and specific object type data). This allows data to be easily swapped between objects when shuffling object data in ObjectState. Logic accessing this data was adjusted to work with the new system in UndoObject, PlayerActionChecks, and ObjectSpriteSwapper

In anticipation of log and rock objects within water, I also added water to the undo system, storing two booleans (one for rock in water, one for log in water)