Log Sinking
log can be pushed into water where it makes a one-time use platform that sinks when the player enters it and then leaves.  added isDisabled state to UndoObject and ObjectState.

Also reworked ObjectSpriteSwapper to be more dynamic and easier to use/expand.

Note some behavior for the log is not done due to sorting order layering issues when a log and water are on the same grid position