Tree prefab (and top sprite handling)
TreeTopEnabler now enables or disables the top asset of the tree based on whether the object is of object type tree. Also the top sprite is always layered in front of the player and other objects on the same panel. Added lower section of tree to ObjectSpriteSwapper (also added bush -> now quantumization works right with both)

Also: adjusted default sorting orders on panel prefabs for better clarity in editor