Compy placing and collecting
placing moves compyReference level object. Collecting disables it, restoring the ability charge to place it again.

the reference is initialized as disabled upon start if the player has the compy in their dinoType list

Compy can be configured to be placed in the scene by default - may use this in a couple levels